requirements.txt
alembic==1.7.7
astroid==2.4.2
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clear==1.0.1
click==8.0.4
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.6.0
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
goodlogging==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.5.4
itsdangerous==2.1.0
Jinja2==3.0.3
lazy-object-proxy==1.4.3
Mako==1.2.0
MarkupSafe==2.1.0
mccabe==0.6.1
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pygame==2.0.0
pylint==2.6.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
rarfile==2.7
requests==2.6.0
six==1.15.0
SQLAlchemy==1.4.32
toml==0.10.1
tomli==2.0.1
urllib3==1.26.8
virtualenv==20.13.3
Werkzeug==2.0.3
wrapt==1.12.1
WTForms==3.0.1
zipp==3.7.0
